Tg invest analyz/Parser-app/requirements.txt
emoji==2.7.0
Flask==2.1.0
pandas==1.5.3
psycopg2==2.9.6
python-dotenv==1.0.0
requests==2.27.1
SQLAlchemy==2.0.0b1 # ПОМЕНЯТЬ base!!!!
Telethon==1.27.0
